+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+1.3.0 (UNRELEASED)
+----------------------
+  - Fix functional test suite (@phahn)
+  - Python 3.12 is supported (@phahn)
+  - Improve documentation (@phahn)
+  - Improve PEP-484 type hinting (@phahn)
+  - Fix mouse dragging (@phahn)
+
 1.2.0 (2023-06-06)
 ----------------------
   - fixes for api.shutdown and disconnect raise exceptions, #256